requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
chainer==7.8.1
cupy-cuda101==7.8.0
cycler==0.11.0
Django==4.0.4
django-crispy-forms==1.14.0
fastrlock==0.8
filelock==3.6.0
fonttools==4.33.2
GDAL==3.3.2
h5py==3.6.0
kiwisolver==1.4.2
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
protobuf==3.20.1
pyparsing==3.0.8
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
typing-extensions==4.2.0
